WIP: Move more synchronizers to steps
