+++ src/azure/index.ts
@@ -1,8 +1,5 @@
-import createAzureWebLinker from './createAzureWebLinker';
+import createAzureWebLinker from "./createAzureWebLinker";
 
-export {
-  createAzureWebLinker,
-};
+export { createAzureWebLinker };
 
 export * from "./types";
-export * from "./resource-manager/index";
