+++ src/azure/resource-manager/authenticate.test.ts
@@ -1,30 +1,38 @@
-import { Polly } from "@pollyjs/core";
+import { Recording, setupRecording } from "@jupiterone/integration-sdk/testing";
 
-import polly from "../../../test/helpers/polly";
 import config from "../../../test/integrationInstanceConfig";
 import authenticate from "./authenticate";
 
-let p: Polly;
+let recording: Recording;
 
 afterEach(async () => {
-  await p.stop();
+  await recording.stop();
 });
 
 test("authenticate with subscription matching", async () => {
-  p = polly(__dirname, "authenticate");
+  recording = setupRecording({
+    directory: __dirname,
+    name: "authenticate",
+  });
   const credentials = await authenticate(config);
   expect(credentials).toBeDefined();
 });
 
 test("authenticate with subscription not matching", async () => {
-  p = polly(__dirname, "authenticate bad subscription");
+  recording = setupRecording({
+    directory: __dirname,
+    name: "authenticate bad subscription",
+  });
   await expect(
     authenticate({ ...config, subscriptionId: "junk" }),
   ).rejects.toThrow(/not found in tenant/);
 });
 
 test("authenticate with no subscription", async () => {
-  p = polly(__dirname, "authenticate no subscription");
+  recording = setupRecording({
+    directory: __dirname,
+    name: "authenticate no subscription",
+  });
   await expect(
     authenticate({
       ...config,
