--- src/azure/resource-manager/index.ts
@@ -1,3 +1,0 @@
-import ResourceManagerClient from "./ResourceManagerClient";
-
-export { ResourceManagerClient };
