--- src/converters/index.ts
@@ -1,4 +1,0 @@
-export * from "./graph";
-export * from "./relationships";
-export * from "./resources/compute";
-export * from "./resources/database";
