+++ src/steps/active-directory/__recordings__/iterateDirectoryRoleMembers_226809899/recording.har
