+++ src/steps/active-directory/__recordings__/iterateGroupMembersSelectProperties_431024926/recording.har
