+++ src/steps/resource-manager/compute/client.test.ts
@@ -1,4 +1,3 @@
-import { NetworkInterface } from "@azure/arm-network/esm/models";
 import { createMockIntegrationLogger } from "@jupiterone/integration-sdk/testing";
 
 import {
@@ -6,7 +5,8 @@ import {
   setupAzureRecording,
 } from "../../../../test/helpers/recording";
 import config from "../../../../test/integrationInstanceConfig";
-import { NetworkClient } from "./client";
+import { ComputeClient } from "./client";
+import { VirtualMachine, Disk } from "@azure/arm-compute/esm/models";
 
 let recording: Recording;
 
@@ -14,21 +14,21 @@ afterEach(async () => {
   await recording.stop();
 });
 
-describe("iterateNetworkInterfaces", () => {
+describe("iterateVirtualMachines", () => {
   test("all", async () => {
     recording = setupAzureRecording({
       directory: __dirname,
-      name: "iterateNetworkInterfaces",
+      name: "iterateVirtualMachines",
     });
 
-    const client = new NetworkClient(
+    const client = new ComputeClient(
       config,
       createMockIntegrationLogger(),
       true,
     );
 
-    const resources: NetworkInterface[] = [];
-    await client.iterateNetworkInterfaces((e) => {
+    const resources: VirtualMachine[] = [];
+    await client.iterateVirtualMachines((e) => {
       resources.push(e);
     });
 
@@ -36,6 +36,36 @@ describe("iterateNetworkInterfaces", () => {
       expect.objectContaining({
         id: expect.any(String),
         name: "j1dev",
+        tags: expect.objectContaining({
+          environment: "j1dev",
+        }),
+      }),
+    ]);
+  });
+});
+
+describe("iterateVirtualMachineDisks", () => {
+  test("all", async () => {
+    recording = setupAzureRecording({
+      directory: __dirname,
+      name: "iterateVirtualMachineDisks",
+    });
+
+    const client = new ComputeClient(
+      config,
+      createMockIntegrationLogger(),
+      true,
+    );
+
+    const resources: Disk[] = [];
+    await client.iterateVirtualMachineDisks((e) => {
+      resources.push(e);
+    });
+
+    expect(resources).toEqual([
+      expect.objectContaining({
+        id: expect.any(String),
+        name: "j1devOsDisk",
         tags: expect.objectContaining({
           environment: "j1dev",
         }),
