--- src/utils/getTime.ts
@@ -1,5 +1,0 @@
-export default function getTime(
-  time: Date | string | undefined,
-): number | undefined {
-  return time ? new Date(time).getTime() : undefined;
-}
