+++ src/validateInvocation.test.ts
@@ -1,11 +1,12 @@
 import { IntegrationValidationError } from "@jupiterone/integration-sdk";
 import { createMockExecutionContext } from "@jupiterone/integration-sdk/testing";
 
+import { IntegrationConfig } from "./types";
 import validateInvocation from "./validateInvocation";
 
 it("should reject", async () => {
-  const executionContext = createMockExecutionContext({
-    instanceConfig: {},
+  const executionContext = createMockExecutionContext<IntegrationConfig>({
+    instanceConfig: {} as IntegrationConfig,
   });
 
   try {
